requirements.txt
astroid==2.0.4
autopep8==1.4.2
Django==2.1.5
django-otp==0.5.1
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djoser==1.3.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
